+++ tests/conftest.py
@@ -3,30 +3,28 @@ import pytest
 
 @pytest.fixture
 def create_accounts():
-    from tests.fixtures.accounts import create_accounts as _create_accounts
+    from tests.fixtures.account import create_accounts as _create_accounts
 
     return _create_accounts
 
 
 @pytest.fixture
 def create_trading_account():
-    from tests.fixtures.accounts import (
-        create_trading_account as _create_trading_account,
-    )
+    from tests.fixtures.account import create_trading_account as _create_trading_account
 
     return _create_trading_account
 
 
 @pytest.fixture
 def btc_usd_market_json_data():
-    from tests.fixtures.markets import get_btc_usd_market_json_data
+    from tests.fixtures.market import get_btc_usd_market_json_data
 
     return get_btc_usd_market_json_data()
 
 
 @pytest.fixture
 def create_btc_usd_market(btc_usd_market_json_data):
-    from tests.fixtures.markets import create_btc_usd_market as _create_btc_usd_market
+    from tests.fixtures.market import create_btc_usd_market as _create_btc_usd_market
 
     return lambda: _create_btc_usd_market(btc_usd_market_json_data)
 
@@ -42,7 +40,7 @@ def create_orderbook_message():
 
 @pytest.fixture
 def create_account_update_trade_message():
-    from tests.fixtures.accounts import (
+    from tests.fixtures.account import (
         create_account_update_trade_message as _create_account_update_trade_message,
     )
 
@@ -51,7 +49,7 @@ def create_account_update_trade_message():
 
 @pytest.fixture
 def create_account_update_unknown_message():
-    from tests.fixtures.accounts import (
+    from tests.fixtures.account import (
         create_account_update_unknown_message as _create_account_update_unknown_message,
     )
 
@@ -60,22 +58,20 @@ def create_account_update_unknown_message():
 
 @pytest.fixture
 def get_asset_usd():
-    from tests.fixtures.assets import get_asset_usd as _get_asset_usd
+    from tests.fixtures.asset import get_asset_usd as _get_asset_usd
 
     return _get_asset_usd
 
 
 @pytest.fixture
 def get_asset_xvs():
-    from tests.fixtures.assets import get_asset_xvs as _get_asset_xvs
+    from tests.fixtures.asset import get_asset_xvs as _get_asset_xvs
 
     return _get_asset_xvs
 
 
 @pytest.fixture
 def create_asset_operations():
-    from tests.fixtures.assets import (
-        create_asset_operations as _create_asset_operations,
-    )
+    from tests.fixtures.asset import create_asset_operations as _create_asset_operations
 
     return _create_asset_operations
