+++ tests/fixtures/market.py
@@ -185,7 +185,7 @@ def get_btc_usd_market_json_data():
 
 
 def create_btc_usd_market(json_data: str):
-    from x10.perpetual.markets import MarketModel
+    from x10.models.market import MarketModel
     from x10.utils.http import WrappedApiResponse
 
     result = WrappedApiResponse[List[MarketModel]].model_validate_json(json_data)
