+++ tests/perpetual/order_object/test_order_object_attrs.py
@@ -6,8 +6,8 @@ from freezegun import freeze_time
 from hamcrest import assert_that, equal_to, has_entries
 from pytest_mock import MockerFixture
 
-from x10.perpetual.configuration import TESTNET_CONFIG
-from x10.perpetual.orders import OrderSide
+from x10.configuration import TESTNET_CONFIG
+from x10.models.order import OrderSide
 from x10.utils.date import utc_now
 
 FROZEN_NONCE = 1473459052
