+++ tests/perpetual/test_orderbook_price_impact.py
@@ -2,9 +2,9 @@ import asyncio
 import decimal
 from unittest import TestCase
 
-from x10.perpetual.configuration import TESTNET_CONFIG
+from x10.configuration import TESTNET_CONFIG
+from x10.models.orderbook import OrderbookUpdateModel
 from x10.perpetual.orderbook import OrderBook
-from x10.perpetual.orderbooks import OrderbookUpdateModel
 
 
 class TestOrderBook(TestCase):
