+++ tests/utils/test_model.py
@@ -4,7 +4,7 @@ from typing import Optional
 from hamcrest import assert_that, equal_to, raises
 from pydantic import ValidationError
 
-from x10.utils.model import X10BaseModel
+from x10.models.base import X10BaseModel
 
 
 class _TestModel(X10BaseModel):
